Renamed store to index
